+++ src/configuration/test.rs
@@ -25,4 +25,12 @@ impl Configuration for TestConfiguration {
     fn default_tags(&self) -> Vec<String> {
         vec![]
     }
+
+    fn socket_path(&self) -> Option<String> {
+        None
+    }
+
+    fn batching_options(&self) -> Option<dogstatsd::BatchingOptions> {
+        None
+    }
 }